Avoid duplicated code for session loading/saving for Matplotlib viewers
