+++ glue/viewers/scatter/qt/data_viewer.py
@@ -12,9 +12,7 @@ from glue.viewers.scatter.qt.layer_style_editor import ScatterLayerStyleEditor
 from glue.viewers.scatter.layer_artist import ScatterLayerArtist
 from glue.viewers.scatter.qt.options_widget import ScatterOptionsWidget
 from glue.viewers.scatter.state import ScatterViewerState
-from glue.viewers.scatter.compat import update_viewer_state
-
-from glue.core.state import lookup_class_with_patches
+from glue.viewers.scatter.compat import update_scatter_viewer_state
 
 __all__ = ['ScatterViewer']
 
@@ -28,6 +26,8 @@ class ScatterViewer(MatplotlibDataViewer):
     _options_cls = ScatterOptionsWidget
     _data_artist_cls = ScatterLayerArtist
     _subset_artist_cls = ScatterLayerArtist
+
+    update_viewer_state = update_scatter_viewer_state
 
     tools = ['select:rectangle', 'select:xrange',
              'select:yrange', 'select:circle',
@@ -90,38 +90,3 @@ class ScatterViewer(MatplotlibDataViewer):
 
             mode = EditSubsetMode()
             mode.update(self._data, subset_state, focus_data=layer_artist.layer)
-
-    def __gluestate__(self, context):
-        return dict(state=self.state.__gluestate__(context),
-                    session=context.id(self._session),
-                    size=self.viewer_size,
-                    pos=self.position,
-                    layers=list(map(context.do, self.layers)),
-                    _protocol=1)
-
-    @classmethod
-    @defer_draw
-    def __setgluestate__(cls, rec, context):
-
-        if rec.get('_protocol', 0) < 1:
-            update_viewer_state(rec, context)
-
-        session = context.object(rec['session'])
-        viewer = cls(session)
-        viewer.register_to_hub(session.hub)
-        viewer.viewer_size = rec['size']
-        x, y = rec['pos']
-        viewer.move(x=x, y=y)
-
-        viewer_state = ScatterViewerState.__setgluestate__(rec['state'], context)
-
-        viewer.state.update_from_state(viewer_state)
-
-        # Restore layer artists
-        for l in rec['layers']:
-            cls = lookup_class_with_patches(l.pop('_type'))
-            layer_state = context.object(l['state'])
-            layer_artist = cls(viewer.axes, viewer.state, layer_state=layer_state)
-            viewer._layer_artist_container.append(layer_artist)
-
-        return viewer
